improve slow tests
