+++ tests/conftest.py
@@ -220,6 +220,7 @@ def training_output_s3_uri(bucket):
 
 
 @pytest.fixture
+@pytest.mark.slow
 def training_job_name(sagemaker_boto_client, training_role_arn, docker_image, training_s3_uri, training_output_s3_uri):
     training_job_name = name()
     sagemaker_boto_client.create_training_job(
@@ -258,6 +259,7 @@ def processing_job_name(sagemaker_boto_client, training_role_arn, docker_image):
 
 
 @pytest.fixture(scope="session")
+@pytest.mark.slow
 def docker_image(boto_model_file):
     client = docker.from_env()
     ecr_client = boto3.client("ecr")
