+++ tox.ini
@@ -53,7 +53,7 @@ passenv =
 # {posargs} can be passed in by additional arguments specified when invoking tox.
 # Can be used to specify which tests to run, e.g.: tox -- -s
 commands =
-    coverage run --source smexperiments -m pytest {posargs}
+    coverage run --source smexperiments -m pytest {posargs} -m "not slow"
     {env:IGNORE_COVERAGE:} coverage report  --fail-under=80
 extras = test
 deps =
@@ -91,3 +91,22 @@ deps =
     pylint
 commands =
     python -m pylint --exit-zero --rcfile=.pylintrc -j 0 src/smexperiments
+
+[testenv:slow-tests]
+passenv =
+    AWS_ACCESS_KEY_ID
+    AWS_SECRET_ACCESS_KEY
+    AWS_SESSION_TOKEN
+    AWS_CONTAINER_CREDENTIALS_RELATIVE_URI
+    AWS_DEFAULT_REGION
+    SAGEMAKER_ENDPOINT
+
+# {posargs} can be passed in by additional arguments specified when invoking tox.
+# Can be used to specify which tests to run, e.g.: tox -- -s
+commands =
+    pytest {posargs} --verbose -m "slow" --capture=no
+extras = test
+deps =
+    boto3 >= 1.10.32
+    pytest
+    docker
